backend/flask_engine/requirements.txt
Flask>=2.2
pandas>=2.0
numpy>=1.24
python-dotenv>=1.0
